Store humidity as decimal

+++ server/src/main/kotlin/dev/drzepka/tempmonitor/server/domain/service/measurement/TestMeasurementDataGenerator.kt
@@ -27,7 +27,7 @@ class TestMeasurementDataGenerator(
     private fun createRandomMeasurement(): MeasurementResource = MeasurementResource().apply {
         time = currentTime
         temperature = BigDecimal.valueOf((random.nextInt(200) + 50) / 10.0).setScale(1, RoundingMode.FLOOR)
-        humidity = random.nextInt(100) + 1
+        humidity = BigDecimal.valueOf(random.nextInt(100) + 1L)
     }
 
 }
